move field type to its own file
